moved parse to another file
